refactor: CarName, CarPosition 객체 생성
