+++ src/main/java/step1/racingcar/domain/vo/CarName.java
@@ -1,0 +1,17 @@
+package step1.racingcar.domain.vo;
+
+import step1.racingcar.exception.CarNameException;
+
+public record CarName(String name) {
+
+    public CarName {
+        validateCarName(name);
+    }
+
+    private static void validateCarName(String name) {
+        if (name.isBlank() || name.length() > 5) {
+            throw new CarNameException("자동차 이름이 5자리 초과되었습니다.");
+        }
+    }
+
+}
