+++ src/test/java/step1/racingcar/RacingCarTest.java
@@ -20,7 +20,6 @@ public class RacingCarTest {
     @DisplayName("자동차 게임 - 1. 자동차 생성 확인")
     public void carCreateTest(int position, boolean expected) throws Exception {
         Car car = new Car("name");
-        car.initializePosition();
         assertThat(car.getPosition() == position).isEqualTo(expected);
     }
 
@@ -28,8 +27,9 @@ public class RacingCarTest {
     @DisplayName("자동차 게임 - 2. 4 이상이면 자동차 전진, 4 미만이면 정지 확인")
     public void carMoveTest() throws Exception {
         Car movingCar = new Car("name");
+        int beforeMovePosition = movingCar.getPosition();
         movingCar.move(new MovableNumberGenerator().generateNumber());
-        assertThat(movingCar.getPosition()).isEqualTo(movingCar.getPosition() + 1);
+        assertThat(movingCar.getPosition()).isEqualTo(beforeMovePosition + 1);
 
         Car nonMovingCar = new Car("name");
         nonMovingCar.move(new ImmovableNumberGenerator().generateNumber());
